requirements.txt
pygame
PyOpenGLp